requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
SQLAlchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
prometheus-client>=0.20.0
aiokafka>=0.11.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
playwright>=1.49.0
